feat(timepicker): create nested component:
