+++ core/components/organisms/timePicker/__stories__/12HourFormat.story.jsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import { TimePicker } from '@/index';
 import { action } from '@/utils/action';
-import { TimePickerAsDropdown, TimePickerAsInput } from './_common_/types';
+import { TimePickerAsDropdown } from '../TimePickerAsDropdown';
+import { TimePickerAsInput } from '../TimePickerAsInput';
 
 // CSF format story
 export const TimePicker12HourFormat = () => {
@@ -50,15 +51,13 @@ const customCode = `() => {
 
 export default {
   title: 'Components/TimePicker/Time Picker 12 Hour Format',
-  component: TimePicker,
+  component: TimePickerAsInput,
+  subcomponents: { TimePickerAsDropdown },
   parameters: {
     docs: {
       docPage: {
         customCode,
         title: 'TimePicker',
-        props: {
-          components: { TimePickerAsInput, TimePickerAsDropdown },
-        },
       },
     },
   },
