--- core/components/organisms/timePicker/__stories__/_common_/types.tsx
@@ -1,6 +1,0 @@
-// TODO(eslint): Fix this
-import { TimePickerWithDropdown, TimePickerWithInput } from '@/components/organisms/timePicker';
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export const TimePickerAsDropdown = (_props: TimePickerWithDropdown) => {};
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export const TimePickerAsInput = (_props: TimePickerWithInput) => {};
